Fixed most issues related to storing the thumbnail on server. One more bug in the Thumbnail URl ...
