+++ mediabox/src/main/js/upload/Upload.js
@@ -144,7 +144,6 @@ class App extends React.Component {
 	this.fileChange = this.fileChange.bind(this);
 	this.imgFileChange = this.imgFileChange.bind(this);
 	this.handleSubmit = this.handleSubmit.bind(this);
-	this.thumbnailUploader = this.thumbnailUploader.bind(this);
     }
 
     //handles a change in an input from the form and gives that new change to the state.
@@ -164,14 +163,35 @@ class App extends React.Component {
     	    this.state.videoThumbnail = files[0];
     }
     
-    thumbnailUploader(){
-    	// initialize thumbnailURL with default value
+    //handles getting state data and giving it to the ajax submit.
+    handleSubmit(e){
+    	e.preventDefault();
+    	
+    	// initialize a "global" thumbnailURL with default value that each post request can access
     	var thumbnailURL = this.state.formData.thumbnailurl;
+    	var torrentF = "";
+
+    	const localForm = new FormData();
+    	localForm.append('video', this.state.videoFile);
+    	
+    	// Start upload process on local mediabox server.
+    	$.post({
+    		url: '/process/upload',
+    		data: localForm,
+    		processData: false,  // tell jQuery not to process the data (because of the file)
+    		contentType: false,  // tell jQuery not to set contentType
+    		success: (torrentFile) => {
+    			torrentF = torrentFile;
+    		},
+    		error: (response) => {
+    			console.log(response);
+    		}
+    	});
     	
     	// make FormData object to store thumbnail image to be uploaded
     	const newForm = new FormData();
-    	newForm.append('video', this.state.videoThumbnail);
-    	
+    	newForm.append('image', this.state.videoThumbnail);
+
     	// post request to upload thumbnail to server
         $.post({
     		url: config.serverUrl + '/upload/thumbnailSubmission',
@@ -182,6 +202,7 @@ class App extends React.Component {
     			// change thumnailURl, if it was uploaded to server, else, leave it as default value
     			if(thumbnailURL_returned !== ""){
     				thumbnailURL = thumbnailURL_returned;
+    				console.log("Thumbnail Submission returned an actual THUMBNAIL URL: " + thumbnailURL_returned);
     			} else {
     				console.log("Thumbnail Submission returned empty String!");
     			}
@@ -190,60 +211,38 @@ class App extends React.Component {
     			console.log("Thumbnail Submission DID NOT WORK!");
     		}
         });
-        return thumbnailURL;
-    }
-    
-    //handles getting state data and giving it to the ajax submit.
-    handleSubmit(e){
-    	e.preventDefault();
+    	
+    	console.log("THUMBNAIL URL = " + thumbnailURL);
+		
+		// ajax request to send video metadata to server
+		// TODO: loop through formData <-- someone else's comment .. not sure why this is needed
+		const formData = {
+				title: this.state.formData.title,
+				description: this.state.formData.description,
+				version: this.state.formData.version,
+				filetype: this.state.videoFile.type,
+				license: this.state.formData.license,
+				downloadurl: torrentF,
+				thumbnailurl: thumbnailURL,
+				tags: this.state.formData.tags,
+				userid: this.state.formData.userid
+		};
 
-    	const localForm = new FormData();
-    	localForm.append('video', this.state.videoFile);
-
-    	// Start upload process on local mediabox server.
-    	$.post({
-    		url: '/process/upload',
-    		data: localForm,
-    		processData: false,  // tell jQuery not to process the data (because of the file)
-    		contentType: false,  // tell jQuery not to set contentType
-    		success: (torrentFile) => {
-    			// have to do this in a separate method, otherwise Cross-Origin Request Blocked
-    			var thumbnailURL = this.thumbnailUploader();
-    			
-    			// ajax request to send video metadata to server
-    			// TODO: loop through formData <-- someone else's comment .. not sure why this is needed
-    			const formData = {
-    					title: this.state.formData.title,
-    					description: this.state.formData.description,
-    					version: this.state.formData.version,
-    					filetype: this.state.videoFile.type,
-    					license: this.state.formData.license,
-    					downloadurl: torrentFile,
-    					thumbnailurl: thumbnailURL,
-    					tags: this.state.formData.tags,
-    					userid: this.state.formData.userid
-    			};
-
-    			// insert video
-    			$.ajax({
-    				type: "POST",
-    				url: config.serverUrl + "/upload/videoSubmission",
-    				contentType: 'application/json',
-    				processData: false,
-    				data: JSON.stringify(formData),
-    				success: (response) => {
-    					console.log(response);
-    					alert("Successfully uploaded!");
-    				},
-    				error: (response) => {
-    					console.log(response);
-    				}
-    			});
-    		},
-    		error: (response) => {
-    			console.log(response);
-    		}
-    	});
+		// insert video
+		$.ajax({
+			type: "POST",
+			url: config.serverUrl + "/upload/videoSubmission",
+			contentType: 'application/json',
+			processData: false,
+			data: JSON.stringify(formData),
+			success: (response) => {
+				console.log(response);
+				alert("Successfully uploaded!");
+			},
+			error: (response) => {
+				console.log(response);
+			}
+		});
     }
 
     render(){
