+++ server/src/main/java/spring/storage/FileSystemStorageService.java
@@ -8,13 +8,18 @@ import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.stream.Stream;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
 import org.springframework.util.FileSystemUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import spring.PostController;
+
 public class FileSystemStorageService implements StorageService {
+    private static final Logger log = LoggerFactory.getLogger(FileSystemStorageService.class);
     private final Path rootLocation;
 
     public FileSystemStorageService(StorageProperties properties) {
@@ -32,9 +37,11 @@ public class FileSystemStorageService implements StorageService {
 	String filename = StringUtils.cleanPath(file.getOriginalFilename());
 	try {
 	    if (file.isEmpty()) {
-		throw new StorageException("Failed to store empty file " + filename);
+	        log.error("Failed to store empty file.");
+	        throw new StorageException("Failed to store empty file " + filename);
 	    }
 	    if (filename.contains("..")) {
+	        log.error("Cannot store file with relative path outside current directory");
 		// This is a security check
 		throw new StorageException(
 		    "Cannot store file with relative path outside current directory "
@@ -45,6 +52,7 @@ public class FileSystemStorageService implements StorageService {
 	}
 	catch (IOException e) {
 	    // System.out.println("**" + e.getMessage() + e.toString());
+	    log.error("Failed to store file " + filename + "..." + e.getMessage() + e.toString());
 	    throw new StorageException("Failed to store file " + filename, e);
 	}
     }
